python/worker/requirements.txt
# Python Worker Module Dependencies

# Core worker dependencies
redis>=5.0.0
supabase>=2.0.0
python-dotenv>=1.0.0

# Phase 2: Video Processing
yt-dlp>=2024.0.0
ffmpeg-python>=0.2.0
requests>=2.31.0

# Phase 2: AI/ML
google-generativeai>=0.3.0
mediapipe>=0.10.0
opencv-python>=4.8.0
numpy>=1.24.0

# Type checking
mypy>=1.0.0
